se separan las asingaciones en proyectos y personas para disminuir la longitud de la funcion, falta lo mismo en administradores
